Add the cannon and other features that are used with it
We now have a list of projectiles and remove projectiles when they move
into a deadzon on the left, right and bottom of the screen. Some of the
physics constants have been tweaked, and I've added air friction.
The cannon follows the cursor and shoots upon mouse click.

This commit will resolve #2 and resolve #3.